Some test restructure
